Update on-channeling-your-natural-genius-zone.md
updated inline links; came up with convention id#-citation|link|note. I define all the links at the bottom of the page using the convention id#-link. This allows me to simply put [link-text][id#-link] in the body of the post and not have any URLs appear in the body of the post markdown.